# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "RevealSecretOut": {
        "description": "The decrypted secret value (admin-only, step-up-MFA-gated reveal).",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "title": "RevealSecretOut",
        "type": "object"
      },
...
      "SetSecretRequest": {
        "description": "Store a free-form named secret (a secret-backend reference value), encrypted at rest.",
        "properties": {
          "ref": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Ref",
            "type": "string"
          },
          "value": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "value"
        ],
        "title": "SetSecretRequest",
        "type": "object"
      },
      "SetSettingRequest": {
        "description": "Set or update one setting's in-app override (in-app value wins over the env default).",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          }
        },
        "required": [
          "value"
        ],
        "title": "SetSettingRequest",
        "type": "object"
      },
      "SettingMutationResult": {
        "description": "Acknowledgement of a settings mutation + whether a restart is needed for full effect.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "overridden": {
            "title": "Overridden",
            "type": "boolean"
          },
          "restart_required": {
            "title": "Restart Required",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "overridden",
          "restart_required",
          "version"
        ],
        "title": "SettingMutationResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API for Fathomline (built on the Fathom engine). The read/query routes are RBAC- and scope-gated; the agent ingest routes authenticate by mTLS client-cert fingerprint behind the proxy; the remediation, preview and deployment surfaces are default-OFF and individually gated. See the project docs for auth and deployment.",
    "summary": "The storage-estate analyzer — read/query surface, agent ingest, and the default-OFF write/deploy surfaces.",
    "title": "Fathomline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings/secrets": {
      "put": {
        "description": "Store a free-form named secret (encrypted at rest) the secret provider resolves by name.",
        "operationId": "set_named_secret_api_v1_settings_secrets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Named Secret",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/secrets/{ref}": {
      "delete": {
        "description": "Delete a stored named secret. 404 if no such secret is set.",
        "operationId": "clear_named_secret_api_v1_settings_secrets__ref__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Named Secret",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/{key}": {
      "delete": {
        "description": "Reset a setting to its env/default value (delete the override). 404 if none was set.",
        "operationId": "clear_setting_api_v1_settings__key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Setting",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "description": "Set or update one setting's in-app override (validated; in-app value wins).\n\nNon-secret by classification, but an *egress-endpoint* key additionally requires fresh step-up\nMFA (``_require_step_up_for_egress``) — changing where a credential is sent is as sensitive as\nrevealing it.",
        "operationId": "set_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSettingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Setting",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/{key}/reveal": {
      "post": {
        "description": "Return the decrypted value of a stored secret (admin-only + fresh step-up MFA).",
        "operationId": "reveal_secret_api_v1_settings__key__reveal_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealSecretOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reveal Secret",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}